### 2018-02-08 #570 @whedon
[PRE REVIEW]: weathercan: Download and format weather data from Environment and Climate Change Canada
**Submitting author:** @steffilazerte (<a href="http://orcid.org/0000-0002-7690-8360">Stefanie LaZerte</a>)
**Repository:** <a href="https://github.com/ropensci/weathercan" target ="_blank">https://github.com/ropensci/weathercan</a>
**Version:** v0.2.4
