### 2018-02-08 #571 @arfon
👋 @steffilazerte - thanks for your submission to JOSS. I realize your submission has already been reviewed by rOpenSci so this should be quick. That said, it looks like your current paper is tailored towards technical documentation which we try to avoid.

Could you please revise your paper to match more closely with our [requested content/structure](http://joss.theoj.org/about#paper_structure)? i.e.:

- A list of the authors of the software and their affiliations
- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software
- A list of key references including a link to the software archive
- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it

 [10.21105.joss.00571.pdf](https://github.com/openjournals/joss-reviews/files/1707417/10.21105.joss.00571.pdf)
