### 2018-02-08 #571 @steffilazerte
Hi, so essentially you mean to remove the usage aspect of the paper? And keep it mostly descriptive?
