### 2018-02-08 #571 @arfon
> Hi, so essentially you mean to remove the usage aspect of the paper? And keep it mostly descriptive?

Yes please. We believe that the usage content should reside in documentation that is linked to from the paper. Does that make sense?
