### 2018-02-10 e0ba341
Simplify JOSS manuscript
