+++ inst/JOSS/paper.md
@@ -26,142 +26,11 @@ bibliography: paper.bib
 
 Environment and Climate Change Canada maintains an online source of historical Canadian weather data in hourly, daily and monthly formats for various stations across Canada [@canada_historical_2011]. This data is freely available and can be accessed directly from their website. However, downloading data from multiple stations and across larger time periods can take significant time and effort. Further, these downloads require processing before they can be used for analysis. `weathercan` [@lazerte_weathercan_2018] is an R [@r_stats] package that automates and simplifies the downloading and formating of this data.
 
-The first step in using `weathercan` is to identify the `station_id`(s) of the weather station(s) of interest. Stations can be searched for either by name (e.g., `stations_search("Brandon")`) or proximity to a given location (e.g., `stations_search(coords = c(49.84847, -99.95009))`). Searches can be conducted on all possible stations, or filtered to include only those recording weather at the desired time interval:
+The first step in using `weathercan` is to identify the station ID(s) of the weather station(s) of interest. Stations can be searched for either by name or proximity to a given location. Searches can be conducted on all possible stations, or filtered to include only those recording weather at the desired time interval. Next, weather data can be downloaded for the specified stations, time range and time interval (i.e. hours, days, months). Data downloaded from multiple stations and over several months are automatically combined into one data frame ready for analysis or plotting (Figure 1). Finally, weather data from a single station can be aligned and merged with existing datasets through linear interpolation.
 
-``` {.r}
-library(weathercan)
-stations_search("Brandon", interval = "hour")
-```
-
-    ## # A tibble: 3 x 12
-    ##   prov   station_name station_id clima… WMO_id TC_id   lat    lon  elev inte… start   end
-    ##   <fctr> <chr>        <fctr>     <fctr> <fctr> <fct> <dbl>  <dbl> <dbl> <chr> <int> <int>
-    ## 1 MB     BRANDON A    3471       50104… 71140  YBR    49.9 -100.0   409 hour   1958  2012
-    ## 2 MB     BRANDON A    50821      50104… 71140  YBR    49.9 -100.0   409 hour   2012  2018
-    ## 3 MB     BRANDON RCS  49909      50104… 71136  PBO    49.9 -100.0   409 hour   2012  2018
-
-Next, weather data can be downloaded for the specified stations (`station_ids`), time range (`start` through to `end`) and time interval (`interval`). Data downloaded from multiple stations and over several months are automatically combined into one data frame ready for analysis or plotting (Figure 1).
-
-``` {.r}
-w <- weather_dl(station_ids = c(50821, 51097), 
-                start = "2017-01-01", end = "2017-09-01",
-                interval = "hour")
-names(w)
-```
-
-    ##  [1] "station_name"    "station_id"      "prov"            "lat"            
-    ##  [5] "lon"             "elev"            "climate_id"      "WMO_id"         
-    ##  [9] "TC_id"           "date"            "time"            "year"           
[…]
