### 2018-02-10 #571 @steffilazerte
Okay, I've updated the manuscript, let me know if this should be even further simplified.
